Merging Long's branch into Tin's branch
Merging branches so that Long's popup button would activate Tin's popup, which works though the popup button is still broken due to an accident on Long's side.

To note, this is also a demonstration of merging branches and resolving conflicts between our branches. Presumably not the intended way Datt wanted, will have to update again.